Add some changes to make future handling of command-line arguments easier.
main.py:
In the global 'main' function: Accept a use_gui argument. If the 'use_gui' argument isn't supplied, derive the value from the configuration file. Pass the value to the Mapper and Server threads which require it for creating/destroying the window instance and queue.

mapper.py:
Remove the 'USE_GUI' constant definition as it is no longer being used.
In the 'Mapper' class: accept use_gui as an argument to the constructor, and pass it on to 'World.__init__'.

emulation.py:
Remove the 'USE_GUI' constant definition as it is no longer being used.
In the global 'main' function: Accept a use_gui argument. If the 'use_gui' argument isn't supplied, derive the value from the configuration file. Pass the value to the Emulator thread which requires it for creating/destroying the window instance and queue. Remove creating an instance of 'EmulatedWorld', this will be done inside the 'Emulator' class.
In the 'Emulator' class: Create the instance of 'EmulatedWorld' inside the constructor, rather than requiring a reference to it in the arguments. Add Use_gui as an argument to the constructor, and pass it to the 'EmulatedWorld' instance. Move the definition of 'wld' in the 'run' method outside the while loop for efficiency.
In the 'EmulatedWorld' class: Accept 'use_gui' as an argument to the constructor, and pass it on to 'World.__init__'. Also move the definition of self.width and self.height inside the constructor.