Change the syntax of the version string that the PyInstaller script looks for. The new format is now v[version]-[stable|beta]-?[build]. Note that the '-' before the build number is optional, but the one before stable/beta is not. For example, a valid version string might look like v1.1-stable, v1.1-beta, v1.1-beta5 or v1.1-beta-5.
Update the PyInstaller script to check the command line arguments for a version string to build.
The new order used when searching for a version string is command line arguments, version.ignore file, most recent tag from Git, or v0.0-beta if none of these were found.

Clean up the code for expanding the version number into a CSV string.