Rename World.isExitLogical to World.isBidirectional.
In the high contrast GUI:
rename self.spacer to self.gap
When the mouse moves, highlight rooms under the mouse and speak their names.
Left clicking on a room prints the speed walk directions from the current location to that room.
Left clicking while holding shift prints the vnum of that room.
Right clicking on a room sets the current room to that room.
Moving the scroll wheel increases or decreases the zoom.
Pressing the middle mouse button (often the same as pressing the scroll wheel) resets the zoom.
Fix up and down arrows not being displayed correctly when in continuous view.
In addition to the blinkers, the center room is now marked with a black circle with a white center.
Cleanup and remove duplicate / unused code.