add a hack so when a user command is called through the emu function, the user command thinks the player is in the emulation room
Example: typing "emu rinfo" will give the rinfo of the room being explored through emulation, and not the room that the char is in.
Likewise typing "emu ravoid +" will set the avoid flag on the room being explored through emulation and not the room that the char is in.
This change has no effect on offline emulation