Fix utils.formatDocString to no longer require the docstring to have an initial indent.
If the first, non-empty line of the docstring passed to utils.formatDocString has no initial indent, it will be prefixed by the white space from the subsequent, non-empty line with the least amount of indention before processing.
This fixes an issue where passing the docstring to textwrap.dedent would do nothing if the first line had no indention, because textwrap.dedent uses the indention of the first line to determine the common indention for the subsequent lines.